Create student monitor report

Equivalent to Import/Export => Student => Monitors student file

Usage

rp_report_student_monitor(
  conn = NULL,
  database = Sys.getenv("RP_DATABASE"),
  format = c("long", "wide")
)

Arguments

conn

DBI connection to use, see rp_connect.

database

Name of the default database to connect to. Retrieved from RP_DATABASE environment variable by default but can be overridden here, see details.

format

Character vector of length one specifying how the data should be formatted. If "long" (the default) return separate rows for each weekday/pickup/dropoff. If "wide" return separate columns for each weekday/pickup/dropoff.

Examples

student_monitor <- rp_report_student_monitor()

## data fields:
dplyr::glimpse(student_monitor[0, ])
#> Rows: 0
#> Columns: 13
#> $ RecordID               <int> 
#> $ `Student ID`           <chr> 
#> $ `Last, First`          <chr> 
#> $ `Program Description`  <chr> 
#> $ `INFO WC`              <chr> 
#> $ `INFO NEED MONITOR`    <chr> 
#> $ `School Abbrev`        <chr> 
#> $ `School Anchor Abbrev` <chr> 
#> $ `School Name`          <chr> 
#> $ Day                    <chr> 
#> $ Shift                  <chr> 
#> $ Route                  <chr> 
#> $ Bus                    <chr> 

## example pickup data:
dplyr::glimpse(
dplyr::select(
  student_monitor,
  -dplyr::contains("ID"), -dplyr::contains("Last"), -dplyr::contains("First")))
#> Rows: 207,702
#> Columns: 10
#> $ `Program Description`  <chr> "DR DOOR FULL DAY", "DR DOOR FULL DAY", "DR DOO…
#> $ `INFO WC`              <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ `INFO NEED MONITOR`    <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ `School Abbrev`        <chr> "0335", "0335", "0420", "0420", "0420", "0420",…
#> $ `School Anchor Abbrev` <chr> "0335", "0335", "0420", "0420", "0420", "0420",…
#> $ `School Name`          <chr> "Farr Acad", "Farr Acad", "LANDMARK HS", "LANDM…
#> $ Day                    <chr> NA, NA, "M", "M", "T", "T", "W", "W", "H", "H",…
#> $ Shift                  <chr> "Pickup", "Dropoff", "Pickup", "Dropoff", "Pick…
#> $ Route                  <chr> NA, NA, "04201451", "04200451", "04201451", "04…
#> $ Bus                    <chr> NA, NA, "7D036", "7D036", "7D036", "7D036", "7D…